refactor: move contacts and cases types to HrmTypes module
