+++ packages/types/HrmTypes/index.ts
@@ -13,16 +13,9 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see https://www.gnu.org/licenses/.
  */
-import { CaseSectionRecord, CaseSection } from '@tech-matters/types/HrmTypes';
 
-export { CaseSectionRecord, CaseSection };
-
-export type CaseSectionUpdate = Omit<
-  CaseSection,
-  'sectionId' | 'createdBy' | 'createdAt' | 'eventTimestamp' | 'sectionType'
-> & {
-  eventTimestamp?: string;
-};
-export type NewCaseSection = Omit<CaseSectionUpdate, 'updatedAt' | 'updatedBy'> & {
-  sectionId?: string;
-};
+export * from './Contact';
+export * from './Referral';
+export * from './ConversationMedia';
+export * from './Case';
+export * from './CaseSection';
